pyproject.toml
[project]
name = "pygamece-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce>=2.5.6",
]

[project.optional-dependencies]
dev = []

[tool.pyright]
stubPath = "typings"
typeCheckingMode = "basic"

[tool.pylsp-mypy]
enabled = true

[tool.mypy]
mypy_path = "typings"
